Move canvas to pygame_gui lib.
