+++ src/components/canvas/icon.py
@@ -3,7 +3,7 @@ import os
 import pygame
 
 from ... import pygame_gui as ui
-from .canvas import CanvasComponent
+from ...pygame_gui.components.canvas import CanvasComponent
 from .keypoint import Keypoint
