+++ src/pygame_gui/components/__init__.py
@@ -1,5 +1,7 @@
 __all__ = [
     'Base',
+    'Canvas',
+    'CanvasComponent',
     'CloseButton',
     'IconButton',
     'Label',
@@ -11,6 +13,7 @@ __all__ = [
 
 from .base import Base
 from .button import CloseButton, IconButton, TextButton
+from .canvas import Canvas, CanvasComponent
 from .containers import RectContainer, RoundedRectContainer
 from .label import Label
 from .root import Root
