+++ src/pygame_gui/components/canvas.py
@@ -1,10 +1,12 @@
 import math
 from typing import List, Tuple
 
-from ... import pygame_gui as ui
+from .. import logger, timer
+from .base import Base
+from .containers import RectContainer
 
 
-class CanvasComponent(ui.components.Base):
+class CanvasComponent(Base):
     '''Methods:
     * setCanvasView(scale, view_x, view_y) -> None
     '''
@@ -44,19 +46,19 @@ class CanvasComponent(ui.components.Base):
 
     def addChild(self, child: 'CanvasComponent') -> None:
         if not isinstance(child, CanvasComponent):
-            ui.logger.error("Child must be CanvasComponent.", ValueError, self)
+            logger.error("Child must be CanvasComponent.", ValueError, self)
             return
         child.setCanvasView(self.scale, 0, 0)
         child.redraw()
         super().addChild(child)
 
-class Canvas(ui.components.RectContainer):
+class Canvas(RectContainer):
     def __init__(self, w: int, h: int, x: int, y: int):
         super().__init__(w, h, x, y)
         self._children: List[CanvasComponent]
         self.interactive_when_active = True
 
-        self.smooth_timer = ui.timer.ProgressTimer(0.1, ui.timer.INTERP_POLYN1)
+        self.smooth_timer = timer.ProgressTimer(0.1, timer.INTERP_POLYN1)
 
         self.scale_before = 1.0
         self.scale_dst = 1.0
@@ -110,7 +112,7 @@ class Canvas(ui.components.RectContainer):
 
     def addChild(self, child: CanvasComponent) -> None:
         if not isinstance(child, CanvasComponent):
-            ui.logger.error("Child must be CanvasComponent.", ValueError, self)
+            logger.error("Child must be CanvasComponent.", ValueError, self)
             return
         child.setCanvasView(self.scale_dst, *self.view_dst)
         super().addChild(child)
